Created "analysis-sweep.jl" which takes an "analysis type" as an argument and applies the corresponding script on the data from a sweep of simulations. For example, `julia analysis-sweep.jl richness` will compute (in parallel) the richness of every run and add it to the database "data_analysis.sqlite". Note that in order for the "richness" argument to be successful, it needs a corresponding directory /richness with a script "richness.jl" that computes the richness of a single run.
"plots-analysis-sweep.jl" is similar to the former. It also computes in 
parallel and takes an analysis type as an argument. However, it creates 
plots for the analyzed data. Note that in order for the "richness" 
argument to be successful, it needs a corresponding directory /richness 
with a script "make-richness-plot.jl".

Added "/src/runmany.jl" for parallel processing.